Add files via upload
An Indian financial services company provides debt financing for a variety of products, including two-wheelers, cars, three-wheelers, and tractors.  The Majorly used loan type amongst the customers is Two-wheelers Loan, which requires no collateral or documentation, allowing customers easy access to their financial needs.   The objective of this project is to predict whether a customer will default their Two-wheeler loan or not.